+++ src/main/java/gtPlusPlus/australia/entity/model/ModelAustralianSpider.java
@@ -38,37 +38,37 @@ public class ModelAustralianSpider extends ModelBase {
         byte b0 = 15;
         this.spiderHead = new ModelRenderer(this, 32, 4);
         this.spiderHead.addBox(-4.0F, -4.0F, -8.0F, 8, 8, 8, f);
-        this.spiderHead.setRotationPoint(0.0F, (float) b0, -3.0F);
+        this.spiderHead.setRotationPoint(0.0F, b0, -3.0F);
         this.spiderNeck = new ModelRenderer(this, 0, 0);
         this.spiderNeck.addBox(-3.0F, -3.0F, -3.0F, 6, 6, 6, f);
-        this.spiderNeck.setRotationPoint(0.0F, (float) b0, 0.0F);
+        this.spiderNeck.setRotationPoint(0.0F, b0, 0.0F);
         this.spiderBody = new ModelRenderer(this, 0, 12);
         this.spiderBody.addBox(-5.0F, -4.0F, -6.0F, 10, 8, 12, f);
-        this.spiderBody.setRotationPoint(0.0F, (float) b0, 9.0F);
+        this.spiderBody.setRotationPoint(0.0F, b0, 9.0F);
         this.spiderLeg1 = new ModelRenderer(this, 18, 0);
         this.spiderLeg1.addBox(-15.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg1.setRotationPoint(-4.0F, (float) b0, 2.0F);
+        this.spiderLeg1.setRotationPoint(-4.0F, b0, 2.0F);
         this.spiderLeg2 = new ModelRenderer(this, 18, 0);
         this.spiderLeg2.addBox(-1.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg2.setRotationPoint(4.0F, (float) b0, 2.0F);
+        this.spiderLeg2.setRotationPoint(4.0F, b0, 2.0F);
         this.spiderLeg3 = new ModelRenderer(this, 18, 0);
         this.spiderLeg3.addBox(-15.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg3.setRotationPoint(-4.0F, (float) b0, 1.0F);
+        this.spiderLeg3.setRotationPoint(-4.0F, b0, 1.0F);
         this.spiderLeg4 = new ModelRenderer(this, 18, 0);
         this.spiderLeg4.addBox(-1.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg4.setRotationPoint(4.0F, (float) b0, 1.0F);
+        this.spiderLeg4.setRotationPoint(4.0F, b0, 1.0F);
         this.spiderLeg5 = new ModelRenderer(this, 18, 0);
         this.spiderLeg5.addBox(-15.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg5.setRotationPoint(-4.0F, (float) b0, 0.0F);
+        this.spiderLeg5.setRotationPoint(-4.0F, b0, 0.0F);
         this.spiderLeg6 = new ModelRenderer(this, 18, 0);
         this.spiderLeg6.addBox(-1.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg6.setRotationPoint(4.0F, (float) b0, 0.0F);
+        this.spiderLeg6.setRotationPoint(4.0F, b0, 0.0F);
         this.spiderLeg7 = new ModelRenderer(this, 18, 0);
         this.spiderLeg7.addBox(-15.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg7.setRotationPoint(-4.0F, (float) b0, -1.0F);
+        this.spiderLeg7.setRotationPoint(-4.0F, b0, -1.0F);
         this.spiderLeg8 = new ModelRenderer(this, 18, 0);
         this.spiderLeg8.addBox(-1.0F, -1.0F, -1.0F, 16, 2, 2, f);
-        this.spiderLeg8.setRotationPoint(4.0F, (float) b0, -1.0F);
+        this.spiderLeg8.setRotationPoint(4.0F, b0, -1.0F);
     }
 
     /**
